+++ sdk/appconfiguration/azure-data-appconfiguration/src/main/java/com/azure/data/appconfiguration/implementation/AzureAppConfigurationImpl.java
@@ -1352,8 +1352,8 @@ public final class AzureAppConfigurationImpl {
                         snapshot,
                         ifMatch,
                         ifNoneMatch,
-                        accept,
                         tags,
+                        accept,
                         context)
                 .map(
                         res ->