+++ sdk/appconfiguration/azure-data-appconfiguration/CHANGELOG.md
@@ -5,12 +5,27 @@
 ### Features Added
 
 - Added a new method `listLabels` to support listing labels capabilities.
+- Added a new property `tagsFilter` to `SettingSelector` to support filtering settings by tags.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core` from `1.49.0` to version `1.49.1`.
+- Upgraded `azure-core-http-netty` from `1.15.0` to version `1.15.1`.
+
+## 1.6.1 (2024-05-28)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core` from `1.48.0` to version `1.49.0`.
+- Upgraded `azure-core-http-netty` from `1.14.2` to version `1.15.0`.
 
 ## 1.6.0 (2024-04-09)